Enum wasm_wrappers::TotalSupply

pub enum TotalSupply {
    Lockable,
    Unlimited,
    Fixed,
}

The token supply of a specific token, set on issuance

Variants

Lockable

Can be issued with no limit, but then can be locked to have a fixed supply.

Unlimited

Unlimited supply, no limits except for numeric limits due to u128

Fixed

On issuance, the total number of coins is fixed